chapter7/2-flask-api-calls/requirements.txt
requests
quart
quart-wtforms
Flask-WTF
aiohttp
aiodns
cchardet
